requirements.txt
blinker==1.9.0
boto3==1.35.68
botocore==1.35.68
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jwt==1.3.1
kafka-python-ng
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
